requirement.txt
SpeechRecognition
pyttsx3
PyAudio
wikipedia
streamlit
pipwin
gTTS
google-generativeai



